Move mappers to separate files
