+++ src/main/java/org/opensearch/knn/index/mapper/MethodFieldMapper.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.opensearch.knn.index.mapper;
+
+import org.apache.lucene.document.FieldType;
+import org.opensearch.common.Explicit;
+import org.opensearch.common.Strings;
+import org.opensearch.common.xcontent.XContentFactory;
+import org.opensearch.knn.index.KNNMethodContext;
+import org.opensearch.knn.index.util.KNNEngine;
+
+import java.io.IOException;
+
+import static org.opensearch.knn.common.KNNConstants.DIMENSION;
+import static org.opensearch.knn.common.KNNConstants.KNN_ENGINE;
+import static org.opensearch.knn.common.KNNConstants.PARAMETERS;
+import static org.opensearch.knn.common.KNNConstants.SPACE_TYPE;
+
+/**
+ * Field mapper for method definition in mapping
+ */
+public class MethodFieldMapper extends KNNVectorFieldMapper {
+
+    MethodFieldMapper(
+        String simpleName,
+        KNNVectorFieldType mappedFieldType,
+        MultiFields multiFields,
+        CopyTo copyTo,
+        Explicit<Boolean> ignoreMalformed,
+        boolean stored,
+        boolean hasDocValues,
+        KNNMethodContext knnMethodContext
+    ) {
+
+        super(simpleName, mappedFieldType, multiFields, copyTo, ignoreMalformed, stored, hasDocValues);
+
+        this.knnMethod = knnMethodContext;
+
+        this.fieldType = new FieldType(KNNVectorFieldMapper.Defaults.FIELD_TYPE);
+
+        this.fieldType.putAttribute(DIMENSION, String.valueOf(dimension));
+        this.fieldType.putAttribute(SPACE_TYPE, knnMethodContext.getSpaceType().getValue());
+
+        KNNEngine knnEngine = knnMethodContext.getKnnEngine();
+        this.fieldType.putAttribute(KNN_ENGINE, knnEngine.getName());
+
+        try {
+            this.fieldType.putAttribute(
+                PARAMETERS,
+                Strings.toString(XContentFactory.jsonBuilder().map(knnEngine.getMethodAsMap(knnMethodContext)))
+            );
+        } catch (IOException ioe) {
+            throw new RuntimeException("Unable to create KNNVectorFieldMapper: " + ioe);
+        }
+
+        this.fieldType.freeze();
+    }
+}
